add: Checking login and presence of cookies for the user and adding a session for the user if there is one

+++ app/controllers/Admin.php
@@ -10,19 +10,34 @@ use Libraries\Validator\Validator;
 class Admin extends Controller{
 
     private $adminModel;
+    private $userModel;
+
     private $req;
     private $validator;
 
 
     public function __construct()
     {
-        $this->adminModel = $this->model('Admin');    
+        $this->adminModel = $this->model('Admin');
+        $this->userModel = $this->model('User');
+
         $this->req = new Request();
         $this->validator = new Validator($this->req);
     }
 
     public function index(){
         
+        if (!Auth::isAuthenticated()) {
+            
+            if(Auth::isAuthenticatedCooke()){
+                $data = $this->userModel->getUserDataById(Auth::getDataCooke()[0]);
+                Auth::loginUser(get_object_vars($data));
+                redirect('admin/index');
+            }else{
+            redirect("");
+            }
+
+        }
         Auth::isAuthenticatedAdmin();
 
         $data['users'] = $this->adminModel->getUsers();
@@ -35,8 +50,20 @@ class Admin extends Controller{
     }
 
     public function addPost(){
-        Auth::isAuthenticatedAdmin();
 
+        if (!Auth::isAuthenticated()) {
+            
+            if(Auth::isAuthenticatedCooke()){
+                $data = $this->userModel->getUserDataById(Auth::getDataCooke()[0]);
+                Auth::loginUser(get_object_vars($data));
+                redirect('admin/index');
+            }else{
+            redirect("");
+            }
+
+        }
+        Auth::isAuthenticatedAdmin();
+    
         $data['userData'] = Auth::getLoggedInUser();
 
         if($this->req->isPostMethod()){
@@ -78,6 +105,19 @@ class Admin extends Controller{
     }
 
     public function editPosts($id){
+
+        if (!Auth::isAuthenticated()) {
+            
+            if(Auth::isAuthenticatedCooke()){
+                $data = $this->userModel->getUserDataById(Auth::getDataCooke()[0]);
+                Auth::loginUser(get_object_vars($data));
+                redirect('admin/index');
+            }else{
+            redirect("");
+            }
+
+        }
+
         Auth::isAuthenticatedAdmin();
 
         $data = $this->adminModel->getDataPostById($id);
@@ -88,6 +128,19 @@ class Admin extends Controller{
     }
 
     public function deletePost($id){
+
+        if (!Auth::isAuthenticated()) {
+            
+            if(Auth::isAuthenticatedCooke()){
+                $data = $this->userModel->getUserDataById(Auth::getDataCooke()[0]);
+                Auth::loginUser(get_object_vars($data));
+                redirect('admin/index');
+            }else{
+            redirect("");
+            }
+
+        }
+
         Auth::isAuthenticatedAdmin();
         if($this->adminModel->deletePost($id)){
             flash('deletePost' , ' پست مورد نظر حذف شد. ');
@@ -100,6 +153,19 @@ class Admin extends Controller{
     }
 
     public function activeUser($id){
+
+        if (!Auth::isAuthenticated()) {
+            
+            if(Auth::isAuthenticatedCooke()){
+                $data = $this->userModel->getUserDataById(Auth::getDataCooke()[0]);
+                Auth::loginUser(get_object_vars($data));
+                redirect('admin/index');
+            }else{
+            redirect("");
+            }
+
+        }
+
         Auth::isAuthenticatedAdmin();
 
         if($this->adminModel->activeUser($id)){
@@ -113,6 +179,19 @@ class Admin extends Controller{
     }
 
     public function deleteUser($id){
+
+        if (!Auth::isAuthenticated()) {
+            
+            if(Auth::isAuthenticatedCooke()){
+                $data = $this->userModel->getUserDataById(Auth::getDataCooke()[0]);
+                Auth::loginUser(get_object_vars($data));
+                redirect('admin/index');
+            }else{
+            redirect("");
+            }
+
+        }
+
         Auth::isAuthenticatedAdmin();
 
         if($this->adminModel->deleteUser($id)){
@@ -126,6 +205,19 @@ class Admin extends Controller{
     }
 
     public function deleteComment($id){
+
+        if (!Auth::isAuthenticated()) {
+            
+            if(Auth::isAuthenticatedCooke()){
+                $data = $this->userModel->getUserDataById(Auth::getDataCooke()[0]);
+                Auth::loginUser(get_object_vars($data));
+                redirect('admin/index');
+            }else{
+            redirect("");
+            }
+
+        }
+
         Auth::isAuthenticatedAdmin();
 
         if($this->adminModel->deleteComment($id)){
@@ -138,6 +230,19 @@ class Admin extends Controller{
     }
 
     public function verifyComment($id){
+
+        if (!Auth::isAuthenticated()) {
+            
+            if(Auth::isAuthenticatedCooke()){
+                $data = $this->userModel->getUserDataById(Auth::getDataCooke()[0]);
+                Auth::loginUser(get_object_vars($data));
+                redirect('admin/index');
+            }else{
+            redirect("");
+            }
+
+        }
+
         Auth::isAuthenticatedAdmin();
         
         if($this->adminModel->verifyComment($id)){
